Add activerecord-import to gemfile for speeding up imports
Break importers down to fast (using activerecord-import) and robust (our old method).  Fast is useful for when the data to be imported is for a fresh DB or you know that it has no cross-over with what is already in there.  Robust is for when you don't know that or know you are re-importing some data.